Multiplayer is working, players are synchronized 🥳
Just need to fix the spawned prefab when the server is a client host

+++ Assets/Scripts/GameState/PlayerState.cs
@@ -4,11 +4,11 @@ namespace Assets.Scripts.GameState
 {
     public class PlayerState
     {
-        public readonly int id;
+        public readonly byte id;
         public Vector3 position;
         public Quaternion rotation;
 
-        public PlayerState(int id, Vector3 position, Quaternion rotation)
+        public PlayerState(byte id, Vector3 position, Quaternion rotation)
         {
             this.id = id;
             this.position = position;
